+++ crates/qfin-api/src/lib.rs
@@ -1,0 +1,6 @@
+pub(crate) mod deserializer;
+
+pub mod error;
+pub use error::Error;
+
+pub mod oanda;